+++ docs/changelog.rst
@@ -3,6 +3,7 @@ v3.11 ( TO BE RELEASED)
 ZCC talker method remove devices added osType attribute (by `Sergio Pereira <mailto:spereira@zscaler.com>`_)
 Added the following methods to ZIA Portal Talker:
   list_malwarePolicy, list_virusSpywareSettings, list_advancedUrlFilteringSettings, list_subscriptions, list_cyberRiskScore
+Updated  README.md and zpa auth docstring
 
 v3.10 ( February 2023)
 =========================